# Benchmark and caching optimization for method accessor