--- a/tests/Avalonia.Benchmarks/Data/AccessorBenchmarks.cs
+++ b/tests/Avalonia.Benchmarks/Data/AccessorBenchmarks.cs
@@ -10,7 +10,8 @@ namespace Avalonia.Benchmarks.Data
     [MemoryDiagnoser, InProcess]
     public class PropertyAccessorBenchmarks
     {
-        private readonly InpcPropertyAccessorPlugin _plugin = new InpcPropertyAccessorPlugin();
+        private readonly InpcPropertyAccessorPlugin _inpcPlugin = new InpcPropertyAccessorPlugin();
+        private readonly MethodAccessorPlugin _methodPlugin = new MethodAccessorPlugin();
         private readonly TestObject _targetStrongRef = new TestObject();
         private readonly WeakReference<object> _targetWeakRef;
 
@@ -20,9 +21,27 @@ namespace Avalonia.Benchmarks.Data
         }
 
         [Benchmark]
-        public void InpcAccessor()
+        public void InpcAccessorMatch()
         {
-            _plugin.Start(_targetWeakRef, nameof(TestObject.Test));
+            _inpcPlugin.Match(_targetWeakRef, nameof(TestObject.Test));
+        }
+
+        [Benchmark]
+        public void InpcAccessorStart()
+        {
+            _inpcPlugin.Start(_targetWeakRef, nameof(TestObject.Test));
+        }
+
+        [Benchmark]
+        public void MethodAccessorMatch()
+        {
+            _methodPlugin.Match(_targetWeakRef, nameof(TestObject.Execute));
+        }
+
+        [Benchmark]
+        public void MethodAccessorStart()
+        {
+            _methodPlugin.Start(_targetWeakRef, nameof(TestObject.Execute));
         }
 
         private class TestObject : INotifyPropertyChanged
@@ -43,6 +62,18 @@ namespace Avalonia.Benchmarks.Data
 
                     OnPropertyChanged();
                 }
+            }
+
+            public void Execute()
+            {
+            }
+
+            public void Execute(object p0)
+            {
+            }
+
+            public void Execute(object p0, object p1)
+            {
             }
 
             public event PropertyChangedEventHandler PropertyChanged;