## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Fixes high overhead for matching for methods and reduces some overhead when starting the subscription. I've added extra benchmarks as well.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
|              Method |          Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------:|----------:|----------:|-------:|------:|------:|----------:|
|   InpcAccessorMatch |     0.2393 ns | 0.0082 ns | 0.0077 ns |      - |     - |     - |         - |
|   InpcAccessorStart |    59.5642 ns | 0.3349 ns | 0.3133 ns | 0.0114 |     - |     - |      48 B |
| MethodAccessorMatch |   477.6047 ns | 6.1541 ns | 5.7565 ns | 0.0935 |     - |     - |     392 B |
| MethodAccessorStart | 2,271.3129 ns | 9.4729 ns | 7.9103 ns | 0.1144 |     - |     - |     480 B |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
|              Method |          Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------:|----------:|----------:|-------:|------:|------:|----------:|
|   InpcAccessorMatch |     0.2344 ns | 0.0173 ns | 0.0162 ns |      - |     - |     - |         - |
|   InpcAccessorStart |    63.0089 ns | 0.2951 ns | 0.2760 ns | 0.0114 |     - |     - |      48 B |
| MethodAccessorMatch |    56.7313 ns | 0.4058 ns | 0.3796 ns |      - |     - |     - |         - |
| MethodAccessorStart | 1,811.2358 ns | 7.6727 ns | 7.1770 ns | 0.0305 |     - |     - |     128 B |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
Since lookup is fairly expensive in this case I am resorting to caching. For `Start` 90%+ of the overhead is caused by delegate creation so we can't do much about it unless we compile our bindings.

## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation